Move constants and error into core
Closes #75
