--- src/bpmncwpverify/constants.py
@@ -1,1 +1,0 @@
-NAMESPACES = {"bpmn": "http://www.omg.org/spec/BPMN/20100524/MODEL"}
