+++ src/bpmncwpverify/core/error.py
