+++ test/core/test_expr_interface.py
@@ -1,7 +1,7 @@
 # type: ignore
 from bpmncwpverify.core.expr import ExpressionListener
 from bpmncwpverify.core.state import SymbolTable
-from bpmncwpverify.error import (
+from bpmncwpverify.core.error import (
     ExpressionComputationCompatabilityError,
     ExpressionNegatorError,
     ExpressionRelationCompatabilityError,
