+++ test/visitors/bpmnchecks/test_bpmn_connectivity_visitor.py
@@ -1,5 +1,5 @@
 # type: ignore
-from bpmncwpverify.error import (
+from bpmncwpverify.core.error import (
     BpmnMsgEndEventError,
     BpmnMsgGatewayError,
     BpmnMsgSrcError,
